+++ rtx_package/src/package/pool/tex_expandable_primitives.rs
@@ -316,7 +316,7 @@ LoadDefinitions!(outer_state, {
 // Hmm... I wonder, should getString itself be dealing with escapechar?
 fn escapechar(state: &State) -> String {
   let code: i32 = match state.lookup_register("\\escapechar", Vec::new()) {
-    Some(RegisterValue::Number(v)) => v.value_of() as i32,
+    Some(RegisterValue::Number(v)) => v.value_of(),
     _ => -1,
   };
   if (0..=255).contains(&code) {
